pyproject.toml
[tool.poetry]
name = "avinsight"
version = "0.1.0"
description = ""
authors = ["Bas Ekkelenkamp <basekkelenkamp1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
flask = "^2.2.3"
flask-socketio = "^5.3.3"
psutil = "^5.9.4"
gpustat = "^1.1"
python-hwinfo = "^0.1.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
